refactor: 경험 상세 정보 아이템 로직 분리
